Create exceptions that are specific to client parameters.
Currently we assume an IllegalArgumentException is caused by an invalid argument provided by the client. This may not always be the case. It is therefore necessary to create an exception that is geared towards invalid client values.

The ParameterException will be used to distinguish invalid client errors, from bad argument values that can occur anywhere in the code.