Add code to compensate for the fact that Solr does not have a normal count function.
As it stands Solr only has a count(distinct()) by the name of unique. It does not have a count() function. This is probably due to the fact that facets always return a count value. This value could be used, but it would require far more changes, than just converting the count into a sum().

See issue: https://www.ebi.ac.uk/panda/jira/browse/GOA-2413